Patch (if loc. and org. need to be changed)
Verify if the location and organizer at the specified ID need to be modified. If some of it need to be modified, for the start verify if they already exist in the database. If yes, assign their id to the specified event else create a new location and organizer in the database, then assign new loc_id and organiztor_id to the specified event (using lastrowid function to get loc and organizer new id).
We do not modify any organizer and location data because they can be assigned to several events and we do not want to change location and organizer to all event assign to it.